dv-5 RSS408-164 - reduce number of ERROR messages logged to console during tests.
When the broker re-uses a testcontainer database - it deletes everything by repeatedly looping over the database tables and attempting to delete them.
This method will attempt to delete tables that have FK constraints.
If the table being deleted is pointed at by another table, the delete will fail.
This is NOT a real error, it's a side effect of the 'dumb loop' and should not be logged as an error.
Eventually all the tables will be deleted.